Update deployment scripts for acreBTC and MidasAllocator
