+++ solidity/deploy/66_transfer_ownership_acrebtc.ts
@@ -1,0 +1,33 @@
+import type { HardhatRuntimeEnvironment } from "hardhat/types"
+import type { DeployFunction } from "hardhat-deploy/types"
+
+const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
+  const { getNamedAccounts, deployments } = hre
+  const { deployer, governance } = await getNamedAccounts()
+  const { log } = deployments
+
+  log(`transferring ownership of acreBTC contract to ${governance}`)
+
+  await deployments.execute(
+    "acreBTC",
+    { from: deployer, log: true, waitConfirmations: 1 },
+    "transferOwnership",
+    governance,
+  )
+
+  if (hre.network.name !== "mainnet" && hre.network.name !== "integration") {
+    await deployments.execute(
+      "acreBTC",
+      { from: governance, log: true, waitConfirmations: 1 },
+      "acceptOwnership",
+    )
+  }
+}
+
+export default func
+
+func.tags = ["TransferOwnershipAcreBTC"]
+func.dependencies = ["acreBTC"]
+func.runAtTheEnd = true
+func.skip = async (hre: HardhatRuntimeEnvironment): Promise<boolean> =>
+  Promise.resolve(hre.network.name === "integration")
